@@ -10,8 +10,8 @@
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "firebase": "^9.6.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-firebase-hooks": "^5.0.3",
     "react-router-dom": "^6.2.2",
     "react-scripts": "^2.1.3",
