@@ -9,7 +9,7 @@ import About from "./components/About/About";
 import NotFound from "./components/NotFound/NotFound";
 import Login from './components/Login/Login';
 import SignUp from './components/Login/SignUp/SignUp';
-// import RequireAuth from './components/RequireAuth/RequireAuth';
+import RequireAuth from './components/RequireAuth/RequireAuth';
 
 function App() {
   return (
@@ -19,7 +19,7 @@ function App() {
         <Route path="/" element={<Shop></Shop>}></Route>
         <Route path="/shop" element={<Shop></Shop>}></Route>
         <Route path="/orders" element={<Orders></Orders>}></Route>
-        <Route path="/inventory" element={<Inventory/>}></Route>
+        <Route path="/inventory" element={<RequireAuth><Inventory/></RequireAuth>}></Route>
         <Route path="/about" element={<About></About>}></Route>
         <Route path="/login" element={<Login></Login>}></Route>
         <Route path="/signup" element={<SignUp></SignUp>}></Route>
