@@ -1,3 +1,4 @@
+import { signOut } from 'firebase/auth';
 import React from 'react';
 import { useAuthState } from 'react-firebase-hooks/auth';
 import { Link } from 'react-router-dom';
@@ -5,6 +6,10 @@ import auth from '../../firebase.init';
 import logo from '../../images/Logo.svg'
 import './Header.css'
 const Header = () => {
+    const [user] = useAuthState(auth);
+    if(user){
+        console.log(user)
+    }
 
     return (
         <nav className='header-nav'>
@@ -14,8 +19,10 @@ const Header = () => {
                 <Link to="/orders">Orders</Link>
                 <Link to="/inventory">Inventory</Link>
                 <Link to="/about">About</Link>
-                <Link to="/login">Login</Link>
-
+                {
+                    user ? <button className="logout-btn" onClick={() => signOut(auth)}>Log Out</button> : <Link to="/login">Login</Link>
+                }
+                
             </div>
         </nav>
     );
