@@ -8,9 +8,10 @@ const RequireAuth = ({ children }) => {
     const location = useLocation();
 
     if(!user){
-        return <Navigate to="/login" state={{ from: location }} replace />;
+        // return <Navigate to="/login" state={{ from: location }} replace />;
+        return <Navigate to="/login"/>;        
+
     }
-    
     return children;
 };
 
